# code/backend/metiche/rest/server/list_plan.go
// Code generated by nuzur go-code-gen. DO NOT EDIT.

package server

import (
	"fmt"
	"log"
	"net/http"

	"go.einride.tech/aip/filtering"
	"go.einride.tech/aip/ordering"

	planmodule "github.com/mklfarha/metiche/backend/core/module/plan"
	"github.com/mklfarha/metiche/backend/core/module/plan/types"
	main_entity "github.com/mklfarha/metiche/backend/entity/plan"
)

func (h *handlers) ListPlan(w http.ResponseWriter, r *http.Request) {
	defaultPageSize := int32(10)
	maxPageSize := int32(100)

	lp, err := parseListParams(r, defaultPageSize, maxPageSize)
	if err != nil {
		writeProblem(w, http.StatusBadRequest, err)
		return
	}

	offset, err := parsePageTokenOffset(lp.pageToken)
	if err != nil {
		writeProblem(w, http.StatusBadRequest, fmt.Errorf("invalid page token: %w", err))
		return
	}

	declarations := planDeclarations()
	filter, err := filtering.ParseFilter(lp, declarations)
	if err != nil {
		writeProblem(w, http.StatusBadRequest, fmt.Errorf("invalid filter expression: %w", err))
		return
	}

	orderBy, err := ordering.ParseOrderBy(lp)
	if err != nil {
		writeProblem(w, http.StatusBadRequest, fmt.Errorf("invalid order_by expression: %w", err))
		return
	}

	req := types.ListRequest{
		Offset:                offset,
		PageSize:              lp.GetPageSize(),
		Filter:                filter,
		FilteringDeclarations: declarations,
		EnumValues:            planEnumValues(),
		OrderBy:               orderBy,
		IncludeFields:         lp.includeFields,
		ExcludeFields:         lp.excludeFields,
	}

	var result types.ListResponse
	if lp.skipCache {
		result, err = h.core.Plan().List(r.Context(), req, planmodule.WithSkipCache())
	} else {
		result, err = h.core.Plan().List(r.Context(), req)
	}
	if err != nil {
		// A filter or order_by the clause builder cannot express is the caller's
		// error, not the server's: statusForError turns it into a 400.
		writeProblem(w, statusForError(err), err)
		return
	}

	resp := ListEnvelope[main_entity.Plan]{
		Items: result.Plan,
	}
	if result.HasNextPage {
		resp.NextPageToken = encodePageToken(offset + int64(lp.GetPageSize()))
	}

	writeJSON(w, http.StatusOK, resp)
}

func planDeclarations() *filtering.Declarations {
	declarations, err := filtering.NewDeclarations(
		filtering.DeclareStandardFunctions(),
		filtering.DeclareIdent("true", filtering.TypeBool),
		filtering.DeclareIdent("false", filtering.TypeBool),

		// plan

		filtering.DeclareIdent("id", filtering.TypeString),

		filtering.DeclareIdent("key", filtering.TypeString),

		filtering.DeclareIdent("name", filtering.TypeString),

		filtering.DeclareIdent("description", filtering.TypeString),

		filtering.DeclareIdent("max_concurrent_agents", filtering.TypeInt),

		filtering.DeclareIdent("retention_days", filtering.TypeInt),

		filtering.DeclareIdent("max_members", filtering.TypeInt),

		filtering.DeclareIdent("max_projects", filtering.TypeInt),

		filtering.DeclareIdent("is_instance_default", filtering.TypeBool),

		filtering.DeclareIdent("sort_order", filtering.TypeInt),

		// An enum is filtered by its VALUE over REST — either quoted
		// (status = "active", matched case-insensitively against every spelling
		// in planEnumValues) or as the generated value name ident
		// (status = STATUS_ACTIVE, the same spelling gRPC accepts). Both are
		// strings to the type checker; the number they stand for is resolved
		// from the value table, not from this declaration.
		filtering.DeclareIdent("status", filtering.TypeString),
		filtering.DeclareIdent("RECORD_STATUS_INVALID", filtering.TypeString),
		filtering.DeclareIdent("RECORD_STATUS_ACTIVE", filtering.TypeString),
		filtering.DeclareIdent("RECORD_STATUS_INACTIVE", filtering.TypeString),

		filtering.DeclareIdent("created_at", filtering.TypeTimestamp),

		filtering.DeclareIdent("updated_at", filtering.TypeTimestamp),
	)
	if err != nil {
		log.Printf("error creating declarations for plan: %v", err)
	}
	return declarations
}

// planEnumValues maps each enum filter field to every accepted
// spelling of its values and the number the column stores. REST generates no
// protobuf enum types (an app built with --api rest has no proto package at
// all), so this table — not the filter declarations — is what makes an enum
// filter resolvable here.
func planEnumValues() map[string]map[string]int64 {
	return map[string]map[string]int64{

		"status": {
			"invalid":                0,
			"RECORD_STATUS_INVALID":  0,
			"active":                 1,
			"RECORD_STATUS_ACTIVE":   1,
			"inactive":               2,
			"RECORD_STATUS_INACTIVE": 2,
		},
	}
}