# code/backend/metiche/rest/server/openapi.yaml
...
openapi: 3.1.0
info:
    title: Metiche
    version: 1.0.0
servers:
    - url: http://localhost:8080/v1
paths:
...
    /plans:
        get:
            tags:
                - Plan
            summary: List plans
            operationId: listPlan
            parameters:
                - name: filter
                  in: query
                  required: false
                  schema:
                    type: string
                - name: order_by
                  in: query
                  required: false
                  schema:
                    type: string
                - name: page_size
                  in: query
                  required: false
                  schema:
                    type: integer
                - name: page_token
                  in: query
                  required: false
                  schema:
                    type: string
                - name: include_fields
                  in: query
                  required: false
                  schema:
                    type: string
                - name: exclude_fields
                  in: query
                  required: false
                  schema:
                    type: string
                - name: skip_cache
                  in: query
                  required: false
                  schema:
                    type: boolean
            responses:
                "200":
                    description: List of plans
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/PlanList'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
        post:
            tags:
                - Plan
            summary: Create Plan
            operationId: createPlan
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Plan'
                required: true
            responses:
                "201":
                    description: Created Plan
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Plan'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "409":
                    description: Conflict
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
...
components:
    schemas:
...
        Plan:
            properties:
                created_at:
                    format: date-time
                    type: string
                description:
                    nullable: true
                    type: string
                id:
                    format: uuid
                    type: string
                is_instance_default:
                    type: boolean
                key:
                    type: string
                max_concurrent_agents:
                    format: int64
                    nullable: true
                    type: integer
                max_members:
                    format: int64
                    nullable: true
                    type: integer
                max_projects:
                    format: int64
                    nullable: true
                    type: integer
                name:
                    type: string
                retention_days:
                    format: int64
                    nullable: true
                    type: integer
                sort_order:
                    format: int64
                    type: integer
                status:
                    $ref: '#/components/schemas/RecordStatus'
                updated_at:
                    format: date-time
                    type: string
            required:
                - id
                - key
                - name
                - is_instance_default
                - sort_order
                - status
                - created_at
                - updated_at
            type: object
        PlanList:
            properties:
                items:
                    items:
                        $ref: '#/components/schemas/Plan'
                    type: array
                next_page_token:
                    type: string
            type: object
...
        Problem:
            properties:
                detail:
                    type: string
                instance:
                    type: string
                status:
                    type: integer
                title:
                    type: string
                type:
                    type: string
            type: object
...
        RecordStatus:
            enum:
                - invalid
                - active
                - inactive
            type: string
...